requirements.txt
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
requests>=2.28.0
beautifulsoup4>=4.11.0
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.12.0 